+++ akshare/index/index_sw.py
@@ -1,10 +1,8 @@
 # -*- coding:utf-8 -*-
 # /usr/bin/env python
 """
-Author: Albert King
-date: 2019/12/6 14:34
-contact: jindaxiang@163.com
-desc: 获取申万指数-申万一级
+Date: 2019/12/6 14:34
+Desc: 获取申万指数-申万一级
 http://www.swsindex.com/IdxMain.aspx
 部分代码要感谢: PKUJson
 """