+++ akshare/nlp/nlp_interface.py
@@ -1,10 +1,8 @@
 # -*- coding:utf-8 -*-
 # /usr/bin/env python
 """
-Author: Albert King
-date: 2020/3/1 0:02
-contact: jindaxiang@163.com
-desc: ownthink-knowledge graph
+Date: 2020/3/1 0:02
+Desc: ownthink-knowledge graph
 https://ownthink.com/
 https://www.ownthink.com/docs/kg/
 """