+++ akshare/qhkc/qhkc_index.py
@@ -1,8 +1,0 @@
-# -*- coding:utf-8 -*-
-# /usr/bin/env python
-"""
-Author: Albert King
-date: 2020/2/14 21:23
-contact: jindaxiang@163.com
-desc: 
-"""