+++ akshare/qhkc/test.py
@@ -1,10 +1,8 @@
 # -*- coding:utf-8 -*-
 # /usr/bin/env python
 """
-Author: Albert King
-date: 2020/2/13 21:33
-contact: jindaxiang@163.com
-desc: 接口测试文件
+Date: 2020/2/13 21:33
+Desc: 接口测试文件
 """
 from akshare.pro.data_pro import pro_api