+++ akshare/utils/token_process.py
@@ -1,10 +1,8 @@
 # -*- coding:utf-8 -*-
 # /usr/bin/env python
 """
-Author: Albert King
-date: 2020/2/13 21:22
-contact: jindaxiang@163.com
-desc: 存储和读取Token文件
+Date: 2020/2/13 21:22
+Desc: 存储和读取Token文件
 """
 import os